feat: release v3 docs
