+++ versioned_docs/version-3.0.0/api-routes-and-mocks.md
@@ -1,0 +1,24 @@
+---
+id: api-routes-and-mocks
+title: Routes and Mocks
+description: mocksServer API
+keywords:
+  - mocks server
+  - api
+  - routes
+  - mocks
+  - properties
+  - options
+---
+
+## Preface
+
+The main concepts to understand when working with Mocks Server are `Routes`, `Route variants` and `Mocks`. That's why their APIs are explained in the "Get started" section.
+
+## Routes
+
+Read the [Routes chapter in the "Get started"](get-started-routes.md) section for a detailed explanation about how to define routes and all of its available properties.
+
+## Mocks
+
+Read the [Mocks chapter in the "Get started"](get-started-mocks.md) section for a detailed explanation about how to define mocks and all of its available properties.
