+++ versioned_docs/version-3.0.0/plugins-admin-api.md
@@ -1,0 +1,49 @@
+---
+id: plugins-admin-api
+title: Administration REST API
+description: How to administrate Mocks Server using the REST API
+keywords:
+  - mocks server
+  - configuration
+  - administration
+  - plugin
+  - api
+  - rest
+  - settings
+  - options
+---
+
+## plugin-admin-api
+
+__The main distribution includes the [plugin-admin-api plugin](https://www.npmjs.com/package/@mocks-server/plugin-admin-api) preinstalled__. It provides a REST API that allows to change the server settings while it is running, etc.
+
+This is __very useful when running acceptance tests, as you can change the current mock__ simply with a HTTP request in the `before` method of your tests, for example.
+
+A __client for the administration api__ is also distributed as a separated package: [@mocks-server/admin-api-client](https://www.npmjs.com/package/@mocks-server/admin-api-client).
+
+## Options
+
+* __`plugins.adminApi.path`__ _(String)_: Path for the administration REST API. Default is `/admin`. You should change it only in case there is any conflict with the API you are mocking.
+
+## API resources
+
+Available API resources are:
+
+* `GET` `/admin/about` Returns "plugin-admin-api" information.
+  * Response body example: `{ "version": "1.2.0" }`
+* `GET` `/admin/mocks` Returns mocks.
+* `GET` `/admin/mocks/:id` Returns a specific mock.
+* `GET` `/admin/routes` Returns routes collection.
+* `GET` `/admin/routes/:id` Returns a specific route.
+* `GET` `/admin/routes-variants` Returns routes variants collection.
+* `GET` `/admin/routes-variants/:id` Returns a specific route variant.
+* `GET` `/admin/mock-custom-routes-variants` Returns an array of currently custom routes variants ids.
+* `POST` `/admin/mock-custom-routes-variants` Defines a route variant to be used by current mock.
+  * Request body example: `{ "id": "users:error" }`
+* `DELETE` `/admin/mock-custom-routes-variants` Restore routes variants to those defined in current mock.
+* `GET` `/admin/settings` Returns current server settings.
+  * Response body example: `{ "server": { "delay": 0 }, "mocks": { "selected": "foo-mock" } }`
+* `PATCH` `/admin/settings` Changes current server settings.
+  * Request body example: `{ "server": { "delay": 3000 } }`
+* `GET` `/admin/alerts` Returns current alerts.
+* `GET` `/admin/alerts/:id` Returns a specific alert. The alert `id` is equivalent to the alert `context` _(read the [developing plugins chapter](plugins-developing-plugins.md) for further info about alerts)_.
