+++ versioned_docs/version-3.0.0/welcome-about-docs.md
@@ -1,0 +1,26 @@
+---
+id: welcome-about-docs
+title: About this documentation
+sidebar_label: About docs
+description: About Mocks Server docs
+keywords:
+  - mocks server
+  - documentation
+  - about
+---
+
+Welcome to the Mocks Server documentation!
+
+## Navigation
+
+You can find different topics in the table of contents. On desktop, you should see it in the left sidebar. On mobile, you should see it after pressing an icon with arrows in the top left corner.
+
+## Something Missing?
+
+If you have ideas for more “How To” recipes that should be on this page, [let us know][issues-url] or [contribute some][repo-docs-url]!
+
+## Feedback
+We are always open to [your feedback][issues-url].
+
+[issues-url]: https://github.com/mocks-server/website/issues
+[repo-docs-url]: https://github.com/mocks-server/website/tree/master/docs/
