+++ versioned_sidebars/version-3.0.0-sidebars.json
@@ -1,0 +1,42 @@
+{
+  "docs": {
+    "Welcome": [
+      "welcome-about-docs"
+    ],
+    "Getting started": [
+      "get-started-intro",
+      "get-started-routes",
+      "get-started-mocks"
+    ],
+    "Guides": [
+      "guides-organizing-files",
+      "guides-using-babel",
+      "guides-proxy-route-variants",
+      "guides-using-middlewares"
+    ],
+    "Releases": [
+      "guides-migrating-from-v2",
+      "guides-migrating-from-v1"
+    ],
+    "Configuration": [
+      "configuration-methods",
+      "configuration-options"
+    ],
+    "Plugins": [
+      "plugins-adding-plugins",
+      "plugins-inquirer-cli",
+      "plugins-admin-api",
+      "plugins-proxy",
+      "plugins-developing-plugins"
+    ],
+    "Integrations": [
+      "integrations-cypress"
+    ],
+    "API": [
+      "api-routes-and-mocks",
+      "api-mocks-server-api",
+      "api-programmatic-usage",
+      "api-routes-handler"
+    ]
+  }
+}
