+++ versions.json
@@ -1,4 +1,5 @@
 [
+  "3.0.0",
   "2.6.0",
   "2.3.0",
   "2.2.0",
